requirements.txt
accelerate==0.21.0
clip-anytorch==2.5.0
deepspeed==0.10.0
deepspeed-mii==0.0.3
facenet-pytorch==2.5.3
googletrans==3.0.0
open-clip-torch==2.20.0
pytorch-lightning==1.9.4
rotary-embedding-torch==0.2.1
s3transfer==0.5.2
sentence-transformers==2.2.2
taming-transformers-rom1504==0.0.6
torch==2.0.1
torch-fidelity==0.3.0
torchaudio==2.0.2
torchdata==0.6.1
torchdiffeq==0.2.3
torchkeras==3.9.3
torchmetrics==1.0.1
torchsde==0.2.5
torchsummary==1.5.1
torchvision==0.15.2
transformers==4.31.0
vector-quantize-pytorch==1.7.1
diffusers==0.23.0